Fixes externalID signup, adds cache policy
Worked with Erin to revert code in ExternalIDRegistrationViewController to implement a signup that uses the externalID as the password.  Need to create the external IDs (WITHOUT CREDENTIALS) on Bridge Study Manager first.  Also added in main view controller from CRFHeartRate and implemements the Cache Policy fields in the Bridge Study info